# doc/api.yaml
openapi: 3.0.3
servers:
  - url: "http://localhost:3000"
info:
  title: WASAText
  description: |-
   Connect with your friends effortlessly using WASAText!
   Send and receive messages, whether one-on-one or in groups,   all from
   the convenience of your PC. Enjoy seamless conversations with text or
   GIFs and easily stay in touch through your private chats or group
   discussions.
  version: 1.0.0
security:
  - BearerAuth: []
#18---------------------------------------------------------------------------------------------------
paths:
...
  /session: 
    post:
      tags: ["login"]
      summary: logs in the user
      description: |-
        creates a user if it doesn't exist, and returns identifier
        if it exists, just returns the user identifier
      operationId: doLogin
      requestBody:
        description: user details
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  example: Maria02
                  minLength: 3
                  maxLength: 16
                  pattern: '^[a-zA-Z0-9]+$'
      responses:
        '201':
          description: user login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  identifier:
                    type: string
                    example: "gugugaga555"
        '400': {$ref: "#/components/responses/BadRequest"}
        '500': {$ref: "#/components/responses/InternalServerError"}
...
  /users/{userID}:
    get:
      operationId: getUserProfile
      description: gets user info, including oneself
      responses:
        '200':
          description: 'user info'
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400": { $ref: "#/components/responses/BadRequest" }
        "500": { $ref: "#/components/responses/InternalServerError" }
    parameters:
      - $ref: "#/components/parameters/userID"
...
  /chats:
    get:
      operationId: getMyConversations
      description: returns all chats in reverse chronological order
      responses:
        '200':
          description: list of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Chat"
        '204':
          description: you have no chats :(
        '500': {$ref: "#/components/responses/InternalServerError"}

    post: 
      operationId: makeChat
      description: makes a new chat, returns the chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - members
              properties:
                chatName:
                  type: string
                photo:
                  type: string
                  format: uri
                members:
                  type: array
                  items:
                    type: string
      responses:
        '201':
          description: made a new chat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chat"
          links: 
            getConversation:
              operationId: getConversation
              description: the chat output can be used to look for that chat resource
              parameters:
                chatID: $response.body#/chatID
        '400': {$ref: "#/components/responses/BadRequest"}
        '409':
          description: convo with this user/group exists already
        '500': {$ref: "#/components/responses/InternalServerError"}
...
  /chats/{chatID}/members/me:
    delete:
      operationId: leaveGroup
      summary: |-
        Can work for both chats and groups because the case 
        where you leave a group and it remains with one member
        creates no conflict with self private chats.
      responses:
        '204':
          description: left group
        '400': {$ref: "#/components/responses/BadRequest"}
        '403':
          description: user not in group/chat
        '500': {$ref: "#/components/responses/InternalServerError"}

    parameters:
      - $ref: "#/components/parameters/chatID"
  /chats/{chatID}/members:
    post:
      operationId: addToGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - targetusername
                - groupID
              properties:
                targetusername:
                  type: string
                groupID:
                  type: string
      responses:
        '200':
          description: new member added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chat"
        '400': {$ref: "#/components/responses/BadRequest"}
        '403':
          description: unauthorised to add other members
        "409":
          description: User is already a member of the group
        '500': {$ref: "#/components/responses/InternalServerError"}
    parameters:
      - $ref: "#/components/parameters/chatID"
...
  /chats/{chatID}/messages/{messageID}/forward:
    post:
      operationId: forwardMessage
      description: forwards messageID to destinationChatID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - destinationChatID
              properties:
                destinationChatID:
                  type: string
                  description: ID of the chat the msg is being fwd to
      responses:
        '201':
          description: message forwarded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        '400': {$ref: "#/components/responses/BadRequest"}
        '500': {$ref: "#/components/responses/InternalServerError"}

    parameters:
      - $ref: "#/components/parameters/chatID"
      - $ref: "#/components/parameters/messageID"


    delete:
      operationId: deleteMessage
      responses:
        '204':
          description: msg deleted
        '400': {$ref: "#/components/responses/BadRequest"}
        '403':
          description: user is not the sender of this msg
        '500': {$ref: "#/components/responses/InternalServerError"}
    parameters:
      - $ref: "#/components/parameters/chatID"
      - $ref: "#/components/parameters/messageID"
  /chats/{chatID}/messages/{messageID}/reactions:
    post:
      operationId: commentMessage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Reaction"
      responses:
        '201':
          description: reaction added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reaction"
          links:
            uncommentMessage:
              operationId: uncommentMessage
              description: the output can be used to check whether user has already reacted
              parameters: 
                reactionID: $response.body#/reactionID
                messageID: $request.path.messageID
                chatID: $request.path.chatID
        '400': {$ref: "#/components/responses/BadRequest"}
        '409':
          description: cannot react more than once! #worth deleting as to allow changing reaction
        '403':
          description: user is not in chat 
        '500': {$ref: "#/components/responses/InternalServerError"}
    parameters:
      - $ref: "#/components/parameters/chatID"
      - $ref: "#/components/parameters/messageID"
...
components:
  schemas:
...
    Chat:
      type: object
      required:
        - chatID
        - groupORchat
        - chatName
        - members
        - creationDate
      properties:
        chatID:
          type: string
        groupORchat:
          type: string
          enum:
            - private
            - group
        chatName:
          type: string
          description: username if private, groupname if group
        members:
          type: array
          items:
            type: string
        creationDate:
          type: string
          format: date-time
        photo:
          description: the profile picture of the user you are chatting with, or an image chosen for the group
          type: string
          format: uri
        mostrecentmsg:
          $ref: "#/components/schemas/MessagePreview"
    User:
      title: user
      description: |-
        a generic user: can refer to oneself, or to other persons
        perusing the app from any device.
      type: object
      required:
        - username
        - userID
      properties:
        photo:
          description: the profile picture of the user you are chatting with
          type: string
          format: uri
        username:
          description: unique username, alphanumeric
          type: string
          example: Maria02
          minLength: 3
          maxLength: 16
          readOnly: true
        userID:
          type: string
          example: ggfvg255
    Message:
      title: message
      description: a single message from a user at a specific date-time
      type: object
      required:
        - messageID
        - senderID
        - status
        - timestamp
      oneOf: 
        - required:
          - content
        - required:
          - photo
      properties:
        messageID:
          type: string
        senderID:
          description: userID of the msg sender
          type: string
        content:
          description: |-
            the word contents of the message, can be none if
            sending a picture
          type: string
        photo:
          type: string
          format: uri
        status:
          type: string
          enum:
            - sent
            - delivered
            - read
        timestamp:
          type: string
          format: date-time
        reactions:
          type: array
          items:
            $ref: "#/components/schemas/Reaction"
        replyTo:
          type: string
          description: |-
            must contain the msg ID you're replying to. needed
            for backend to load msg preview!
        fwdFrom:
          type: string
          description: ID of the og msg that was fwd
    MessagePreview:
      title: message
      description: a single message from a user at a specific date-time
      type: object
      required:
        - messageID
        - timestamp
        - contentType
        #can't use content and photo
      properties:
        senderID:
          type: string
        content:
          description: text of msg, absent when msg is img-only
          type: string
        contentType:
          type: string
          enum:
            - text
            - photo
            - text photo
        timestamp:
          type: string
          format: date-time
        messageID:
          type: string
    Reaction:
      description: |-
        an emoticon added by a user (including oneself) at the foot of a message
        feature: one reaction per user per message
      type: object
      required:
        - emoji
        - userID
        - reactionID
      properties:
        userID:
          type: string
        emoji:
          type: string
        reactionID:
          type: string
...
  responses:
...
    BadRequest:
      description: request not compliant with documentation
    InternalServerError:
      description: server had an internal error. More info in server logs
  securitySchemes:
      BearerAuth:
        type: http
        scheme: bearer
  parameters:
    chatID:
      name: chatID
      in: path
      required: true
      schema:
        type: string
        format: uuid
    messageID:
      name: messageID
      in: path
      required: true
      schema:
        type: string
        format: uuid
    userID:
      name: userID
      in: path
      required: true
      schema:
        type: string
        format: uuid
...